+++ src/test/java/dev/crean/dayeight/ChristmasLightsTest.java
@@ -16,7 +16,9 @@ public class ChristmasLightsTest {
         ChristmasLights lights = new ChristmasLights(Input.DAY_EIGHT);
         lights.buildCircuits(1000);
         List<Circuit> largestCircuits = lights.getLargestCircuits(3);
-        long result = largestCircuits.stream().mapToInt(it -> it.getJunctionBoxes().size()).reduce(1, (a, b) -> a * b);
+        long result = largestCircuits.stream()
+                .mapToLong(it -> it.getJunctionBoxes().size())
+                .reduce(1L, (a, b) -> a * b);
         Assertions.assertEquals(115885, result);
     }
